STM32NucleoF401RE_SensitiveVibrationSensör
Detailed explanation of your project:

Your project is an STM32F4 microcontroller application that reads data from a vibration sensor and transmits it over UART. The data from the sensor is converted to a digital value using an ADC (analog-to-digital converter). This digital value is then transmitted using a UART (universal asynchronous receiver-transmitter).

The project consists of the following steps:

Initialization of ADC and UART: In this step, the necessary settings for ADC and UART are made.
Reading data from the vibration sensor: The ADC converts the analog voltage from the vibration sensor to a digital value.
Converting the vibration value to Hertz: The value read from the ADC is converted to a vibration value in Hertz.
Transmission of the vibration value over UART: The vibration value is transmitted using the serial communication port over UART.
The project works as follows:

After the microcontroller is initialized, the ADC and UART are initialized.
The microcontroller uses the ADC to read data from the vibration sensor.
The read value is converted to a vibration value in Hertz.
The vibration value is transmitted using the serial communication port over UART.
The project can be used in the following applications:

Monitoring the vibration of machines and equipment
Earthquake detection systems
Monitoring the vibration of motor vehicles
Advantages of your project:

Your project is a simple application that reads data from a vibration sensor and transmits it over UART.
Your project uses an STM32F4 microcontroller. STM32F4 microcontrollers are powerful and versatile microcontrollers.
Your project uses important peripherals such as ADC and UART. ADC and UART are important peripherals used in many different applications.
Your project is an open source project. Open source projects give developers more flexibility.